Implementation of checkpoint data structure and checkpoint manager.
This is the initial implementation of checkpoint datastructure and
checkpoint manager. The checkpoint manager maintains the list of
checkpoints and provides the cursors for TAP clients and persistence.
This checkpoint manager is allocated per vbucket. The unit test is
also implemented to simulate persistence thread, set threads, and
TAP replication threads.

Change-Id: I2e35cf5db11d860c17794bcc69d75ebff5c871aa