Creating TAP connection instances with the last checkpoints persisted.
This allows TAP clients to pass the last checkpoints, which were
successfully persisted in their database, to the TAP server when TAP
clients connect to the TAP server. In addition, if the connection
between a given TAP client and TAP server is closed, the client's
cursors are all removed from the checkpoint managers on the server.

Change-Id: I2bff1fda38139479c427a6383db2ae0aa1260ed1