Do not interleave the backfill with live mutations.
Backfill is no longer interleaved with live mutations to preserve
the consistency of checkpoints in the slave node.

The backfill process is adapted in the master node as follows:
1) Remember the ID of the current open checkpoint Ci at the beginning
of the backfill.
2) Let the backfill completed and then remember the ID of the current
open checkpoint Cj at the time of backfill completion.
3) Start the tap cursor from Ci.
4) When the cursor reaches to the end of Cj, send the checkpoint end
message of Cj.

In this way, we can guarantee that all items transmitted reflects
all mutations up to Cj.

Change-Id: I193b6ad2aa389389f7a6bf2bb1860bcb2e7c2277